gerrit patch files from gerrit.iotivity.org
In preporation for the move from gerrit to gitlab
I have create a patch file for commits that are still
in review.

For the master branch I only grabed patch files that went back a
little over a year.

I didn't do any release banches older than 1.3-rel since no activity has shown up on the older releases.

Updated the .gitattributes file so the line endings were not
messed with for *.patch files.

Signed-off-by: George Nash <george.nash@intel.com>